# cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.FileNode:
    properties:
      children:
        items:
          $ref: '#/definitions/handler.FileNode'
        type: array
      is_dir:
        type: boolean
      name:
        type: string
      path:
        type: string
      size:
        type: integer
    type: object
...
  model.APIEnvelope:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: MrLuoooooo
    url: https://github.com/MrLuoooooo/MrLuoooooagent
  description: AI Agent 后端系统。让大模型安全操控本地开发环境——读写文件、执行命令、联网搜索、查询股票、检索知识库。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/MrLuoooooo/MrLuoooooagent
  title: GoAgent Pro API
  version: 4.1.2
paths:
...
  /workspace:
    get:
      description: 获取当前 Agent 工作目录路径。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIEnvelope'
      summary: 获取工作目录
      tags:
      - 工作目录
    post:
      consumes:
      - application/json
      description: 设置 Agent 的当前工作目录。路径不存在时会自动创建。
      parameters:
      - description: 新的工作目录路径
        in: body
        name: request
        required: true
        schema:
          properties:
            path:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIEnvelope'
      summary: 设置工作目录
      tags:
      - 工作目录
  /workspace/dir:
    get:
      description: 列出指定目录下的文件和子目录（仅一层，不递归）。
      parameters:
      - description: 目录路径（默认当前工作目录）
        in: query
        name: path
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIEnvelope'
      summary: 列出目录内容
      tags:
      - 工作目录
...
  /workspace/tree:
    get:
      description: 以树形结构返回指定目录的内容（递归，默认深度 3 层）。
      parameters:
      - description: 起始目录路径（默认当前工作目录）
        in: query
        name: path
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.APIEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/handler.FileNode'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIEnvelope'
      summary: 返回目录树
      tags:
      - 工作目录
swagger: "2.0"